requirements.txt
# ==================== PRODUCTION DEPENDENCIES ====================
# Updated: 2026-04-06 — all versions verified compatible with Python 3.13

# --- TELEGRAM BOT ---
python-telegram-bot>=22.7

# --- DATA & ANALYSIS ---
pandas>=3.0.0
pandas_ta>=0.4.0
numpy>=2.2.0,<2.5
numba>=0.61.0

# --- AI & SENTIMENT ---
transformers>=5.4.0
torch>=2.11.0
scipy>=1.17.0
openai>=2.30.0

# --- VISUALIZATION ---
matplotlib>=3.10.0
Pillow>=12.1.0

# --- UTILITIES ---
requests>=2.33.0
python-dotenv>=1.2.0
feedparser>=6.0.10
flask>=3.1.0
tensorflow>=2.20.0
scikit-learn>=1.8.0
xgboost>=3.0.0
yfinance>=1.2.0

# --- REAL-TIME DATA ---
websocket-client>=1.9.0

# --- SIGNAL PROCESSING ---
PyWavelets>=1.8.0

# --- DATABASE ---
# sqlite3 is built-in Python module

# --- AUTH ---
PyJWT>=2.12.0
bcrypt>=4.0.0

# --- MACRO DATA & NEWS ---
fredapi>=0.5.0
finnhub-python>=2.4.0

# --- NEWS SIMILARITY (FAISS) ---
sentence-transformers>=5.0.0
faiss-cpu>=1.10.0

# --- FASTAPI & WEB ---
fastapi>=0.135.0
uvicorn[standard]>=0.44.0
websockets>=16.0
pydantic>=2.12.0
python-multipart>=0.0.24
httpx>=0.28.0

# --- TESTING ---
pytest>=9.0.0
pytest-asyncio>=1.3.0
pytest-cov>=6.0.0

# --- UTILITIES & UI ---
colorama>=0.4.6
tqdm>=4.67.0

# --- DATABASE CLOUD (Optional) ---
libsql-client>=0.3.0

# --- SYSTEM ---
psutil>=6.0.0

# --- GPU ACCELERATION (DirectML — AMD/Intel/NVIDIA on Windows) ---
onnxruntime-directml>=1.24.0
tf2onnx>=1.17.0
onnx>=1.21.0
skl2onnx>=1.20.0
onnxmltools>=1.16.0

# --- GPU ACCELERATION (Optional — NVIDIA CUDA) ---
# CuPy — GPU-accelerated NumPy (auto-detected by src/compute.py):
# pip install cupy-cuda12x        # For CUDA 12.x (NVIDIA)
# pip install cupy-cuda11x        # For CUDA 11.x (NVIDIA)
#
# CUDA GPU (NVIDIA only, or via WSL2):
# Install torch with CUDA from https://pytorch.org/get-started/locally/
# TensorFlow auto-detects CUDA GPU in WSL2 (no extra package needed)
#
# Numba (already in deps) provides JIT compilation for CPU hot loops.
# With NVIDIA GPU, Numba also supports CUDA kernels (auto-detected).
